backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography] 
bcrypt
sqlalchemy 
psycopg2-binary 
qdrant-client 
sentence-transformers 
pymupdf
python-docx

ollama
tqdm
pdfplumber

langchain
langchain-core
langchain-ollama
langchain-community  

langchain-groq
InstructorEmbedding
apscheduler


